ref: seperate OTP component and refactor hooks
